% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coef.R
\name{extract_coef}
\alias{extract_coef}
\alias{extract_coef.cv.glmnet}
\alias{extract_coef.default}
\alias{extract_coef.glmboost}
\title{Standardized coefficient extraction}
\usage{
extract_coef(object)

\method{extract_coef}{default}(object)

\method{extract_coef}{glmboost}(object)

\method{extract_coef}{cv.glmnet}(object)
}
\arguments{
\item{object}{An object for which the extraction of model coefficients is meaningful.}
}
\value{
Model coefficients as a named numeric vector.
}
\description{
Extract model coefficients in a standardized format (as a named numeric vector).
}

