% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_offsets.R
\name{extract_offsets}
\alias{extract_offsets}
\alias{extract_offsets.cv.glmnet}
\alias{extract_offsets.default}
\alias{extract_offsets.glmboost}
\title{Standardized offset extraction}
\usage{
extract_offsets(object)

\method{extract_offsets}{default}(object)

\method{extract_offsets}{glmboost}(object)

\method{extract_offsets}{cv.glmnet}(object)
}
\arguments{
\item{object}{An object for which the extraction of model offsets is meaningful.}
}
\value{
Model offsets as a numeric vector with all values 1 and formula terms as names.
}
\description{
Extract model offsets (if present) in a standardized format. The return value is a numeric vector
with formula terms as names and values all 1.
}

