% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{fqtn}
\alias{fqtn}
\title{Fully qualify and escape a table name given in parts.}
\usage{
fqtn(table, catalog = NULL, schema = NULL, con = NULL)
}
\arguments{
\item{table}{The unqualified table name.}

\item{catalog}{The catalog name.}

\item{schema}{The schema name.}

\item{con}{An optional DBI connection to control the details of SQL generation}
}
\value{
A dplyr SQL object representing the fully qualified and escaped table name.
}
\description{
Given the three canonical parts of a database table name (catalog, schema
and table), escape all three and join them together to produce the fully
qualified table name. Only the table name is required.
}

