% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpred.R
\name{linpred}
\alias{linpred}
\title{Unevaluated prediction expressions for models}
\usage{
linpred(mod)
}
\arguments{
\item{mod}{A model object providing a coef() method.}
}
\value{
An unevaluated R call object representing the linear predictor.
}
\description{
Generate an unevaluated call corresponding to the predict step of the passed
model. The call represents the linear predictor in terms of elementary functions
on the underlying column names. Before translation into SQL, it should have a response
function applied by score_expression (which may be a no-op in the case of the
identity response).
}

