% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpred.R, R/score_expression.R, R/score_expression_glm.R
\name{linpred}
\alias{linpred}
\alias{score_expression}
\alias{score_expression.default}
\alias{score_expression.glm}
\alias{score_expression.lm}
\title{Unevaluated prediction expressions for models}
\usage{
linpred(mod)

score_expression(mod, link = NULL)

\method{score_expression}{default}(mod, link = NULL)

\method{score_expression}{glm}(mod, link = NULL)

\method{score_expression}{lm}(mod, link = NULL)
}
\arguments{
\item{mod}{A model object providing a coef() method.}

\item{link}{The name of a custom link function to apply to the linear predictor.}

\item{mod}{A model object providing a coef() method.}
}
\value{
An unevaluated R call object representing the linear predictor.

An unevaluated R call object representing the link-inverse function of the linear predictor.

None
}
\description{
Generate an unevaluated call corresponding to the predict step of the passed
model. The call represents the linear predictor in terms of elementary functions
on the underlying column names. Before translation into SQL, it should have a link
function applied by score_expression (which may be a no-op in the case of the
identity link).

Generate an unevaluated call corresponding to the predict step of the passed
model. The call represents the link function of the linear predictor in terms
of elementary functions on the underlying column names, and is suitable for
direct translation into SQL.
}

