% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlscore.R
\name{create_statement}
\alias{create_statement}
\title{Generate a CREATE TABLE statement from a model}
\usage{
create_statement(mod, dest_table, src_table, dest_schema = NULL,
  dest_catalog = NULL, src_schema = NULL, src_catalog = NULL,
  drop = FALSE, temporary = FALSE, pk = c("id"), response = NULL,
  con = NULL)
}
\arguments{
\item{mod}{A model object providing a coef() method.}

\item{dest_table}{The unqualified DB name of the destination table.}

\item{src_table}{The unqualified DB name of the source table.}

\item{dest_schema}{The DB schema of the destination table.}

\item{dest_catalog}{The DB catalog of the destination table.}

\item{src_schema}{The DB schema of the source table.}

\item{src_catalog}{The DB catalog of the source table.}

\item{drop}{Whether to generate a DROP TABLE IF EXISTS before the CREATE TABLE.}

\item{temporary}{Whether the destination table should be a temporary table.}

\item{pk}{A vector of primary key column names.}

\item{response}{The name of a custom response function to apply to the linear predictor.}

\item{con}{An optional DBI connection to control the details of SQL generation.}
}
\value{
A dplyr SQL object representing the SELECT statement.
}
\description{
Generate a CREATE TABLE statement to score the passed model on a preexisting
database table. The statement will generate predictions entirely in the
database, with no need to fetch data into R. Models need not be GLMs, but
their prediction steps must consist of applying a response function to
a linear predictor.
}

