% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_expression.R
\name{score_expression}
\alias{score_expression}
\alias{score_expression.bayesglm}
\alias{score_expression.cv.glmnet}
\alias{score_expression.default}
\alias{score_expression.glm}
\alias{score_expression.glmboost}
\alias{score_expression.lm}
\title{Unevaluated prediction expressions for models}
\usage{
score_expression(mod, link = NULL)

\method{score_expression}{default}(mod, link = NULL)

\method{score_expression}{glm}(mod, link = NULL)

\method{score_expression}{lm}(mod, link = NULL)

\method{score_expression}{bayesglm}(mod, link = NULL)

\method{score_expression}{glmboost}(mod, link = NULL)

\method{score_expression}{cv.glmnet}(mod, link = NULL)
}
\arguments{
\item{mod}{A model object providing a coef() method.}

\item{link}{The name of a custom link function to apply to the linear predictor.}
}
\value{
An unevaluated R call object representing the link-inverse function of the linear predictor.

None
}
\description{
Generate an unevaluated call corresponding to the predict step of the passed
model. The call represents the link function of the linear predictor in terms
of elementary functions on the underlying column names, and is suitable for
direct translation into SQL.
}

