% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlscore.R
\name{select_statement}
\alias{select_statement}
\title{Generate a SELECT statement from a model}
\usage{
select_statement(mod, src_table, src_schema = NULL, src_catalog = NULL,
  pk = c("id"), response = NULL, con = NULL)
}
\arguments{
\item{mod}{A model object providing a coef() method.}

\item{src_table}{The unqualified DB name of the source table.}

\item{src_schema}{The DB schema of the source table.}

\item{src_catalog}{The DB catalog of the source table.}

\item{pk}{A vector of primary key column names.}

\item{response}{The name of a custom response function to apply to the linear predictor.}

\item{con}{An optional DBI connection to control the details of SQL generation.}
}
\value{
A dplyr SQL object representing the SELECT statement.
}
\description{
Generate a SELECT statement to score the passed model on a preexisting
database table. The statement will generate predictions entirely in the
database, with no need to fetch data into R. Models need not be GLMs, but
their prediction steps must consist of applying a response function to
a linear predictor.
}

